src/main/java/uk/gov/hmcts/reform/divorce/validationservice/rules/divorce/session/LegalProceedings.java
package uk.gov.hmcts.reform.divorce.validationservice.rules.divorce.session;

import com.deliveredtechnologies.rulebook.annotation.Given;
import com.deliveredtechnologies.rulebook.annotation.Result;
import com.deliveredtechnologies.rulebook.annotation.Rule;
import com.deliveredtechnologies.rulebook.annotation.Then;
import com.deliveredtechnologies.rulebook.annotation.When;
import lombok.Data;
import uk.gov.hmcts.reform.divorce.validationservice.domain.request.DivorceSession;

import java.util.Optional;

@Rule
@Data
public class LegalProceedings {

    private static final String BLANK_SPACE = " ";
    private static final String ACTUAL_DATA = "Actual data is: %s";
    private static final String ERROR_MESSAGE = "legalProceedings can not be null or empty.";

    @Result
    public String result;

    @Given("divorceSession")
    public DivorceSession divorceSession = new DivorceSession();

    @When
    public boolean when() {
        return !Optional.ofNullable(divorceSession.getLegalProceedings()).isPresent();
    }

    @Then
    public void then() {
        result = String.join(
            BLANK_SPACE, // delimiter
            ERROR_MESSAGE,
            String.format(ACTUAL_DATA, divorceSession.getLegalProceedings())
        );
    }
}